Removed 'startup.py' from 'demo_scripts' folder, adjusted remaining code as necessary, adjusted test and made sure they still run and succeed.
