+++ demo_scripts/draw_overlay_advanced.py
@@ -1,4 +1,4 @@
-import Classes.custom_overlay_class
+from demo_scripts.Classes import custom_overlay_class
 from helpers.printing import try_print_except
 import importlib
 import qtm
@@ -13,7 +13,7 @@ def _reload_script_modules():
     # means that changes you make to these scripts will not show up in
     # QTM despite pressing the "Reload scripts" button. However, running
     # "importlib.reload()" on these scripts will force Python to reload them.
-    importlib.reload(Classes.custom_overlay_class)
+    importlib.reload(custom_overlay_class)
 # endregion
 
 
@@ -36,4 +36,4 @@ def update_and_draw_overlay(measurement_time):
 # ////////   E N T R Y   P O I N T (local 'main')   ////////
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 _reload_script_modules()
-custom_overlay_instance = Classes.custom_overlay_class.custom_overlay()
+custom_overlay_instance = custom_overlay_class.custom_overlay()
